Initial upload of collector agent x API server files
These files relate to the initial idea of how to tackle the server monitoring project. The idea was to create an entirely new server monitoring system, with a collector agent that access the performance metrics and connects to the API server which send the data to the DB for storage.